renamed "/cibm" endpoint to "/ibmcompat" endpoint
