+++ app/src/main/java/io/apicurio/registry/ibmcompat/model/SchemaInfo.java
@@ -1,18 +1,21 @@
-package io.apicurio.registry.cibm.model;
+package io.apicurio.registry.ibmcompat.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Objects;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 
 @javax.annotation.Generated(value = "org.openapitools.codegen.languages.JavaResteasyServerCodegen")
-public class SchemaSummary {
+public class SchemaInfo {
 
     private String id;
     private String name;
     private SchemaState state;
-    private boolean enabled;
+    private Boolean enabled;
+    private List<SchemaVersion> versions = new ArrayList<SchemaVersion>();
 
     /**
      * Lower-case URL-encoded version of the schema name.
@@ -63,12 +66,26 @@ public class SchemaSummary {
 
     @JsonProperty("enabled")
     @NotNull
-    public boolean isEnabled() {
+    public Boolean getEnabled() {
         return enabled;
     }
 
-    public void setEnabled(boolean enabled) {
+    public void setEnabled(Boolean enabled) {
         this.enabled = enabled;
+    }
+
+    /**
+     *
+     **/
+
+    @JsonProperty("versions")
+    @NotNull
+    public List<SchemaVersion> getVersions() {
+        return versions;
+    }
+
+    public void setVersions(List<SchemaVersion> versions) {
+        this.versions = versions;
     }
 
 
@@ -80,27 +97,29 @@ public class SchemaSummary {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        SchemaSummary schemaSummary = (SchemaSummary) o;
-        return Objects.equals(id, schemaSummary.id) &&
-               Objects.equals(name, schemaSummary.name) &&
-               Objects.equals(state, schemaSummary.state) &&
-               Objects.equals(enabled, schemaSummary.enabled);
+        SchemaInfo schemaInfo = (SchemaInfo) o;
+        return Objects.equals(id, schemaInfo.id) &&
+               Objects.equals(name, schemaInfo.name) &&
+               Objects.equals(state, schemaInfo.state) &&
+               Objects.equals(enabled, schemaInfo.enabled) &&
+               Objects.equals(versions, schemaInfo.versions);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(id, name, state, enabled);
+        return Objects.hash(id, name, state, enabled, versions);
     }
 
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append("class SchemaSummary {\n");
+        sb.append("class SchemaInfo {\n");
 
         sb.append("    id: ").append(toIndentedString(id)).append("\n");
         sb.append("    name: ").append(toIndentedString(name)).append("\n");
         sb.append("    state: ").append(toIndentedString(state)).append("\n");
         sb.append("    enabled: ").append(toIndentedString(enabled)).append("\n");
+        sb.append("    versions: ").append(toIndentedString(versions)).append("\n");
         sb.append("}");
         return sb.toString();
     }
@@ -109,7 +128,7 @@ public class SchemaSummary {
      * Convert the given object to string with each line indented by 4 spaces
      * (except the first line).
      */
-    protected String toIndentedString(java.lang.Object o) {
+    private String toIndentedString(java.lang.Object o) {
         if (o == null) {
             return "null";
         }
